+++ src/main/java/io/redstudioragnarok/redcore/vectors/Vector2F.java
@@ -1,5 +1,7 @@
 package io.redstudioragnarok.redcore.vectors;
 
+import io.netty.buffer.ByteBuf;
+
 /**
  * Represents a two-dimensional vector with `x` and `y` coordinates using float.
  *
@@ -28,5 +30,25 @@ public class Vector2F {
     public Vector2F(final float inputX, final float inputY) {
         x = inputX;
         y = inputY;
+    }
+
+    /**
+     * Writes the x and y coordinates of this vector to the given `byteBuf`.
+     *
+     * @param byteBuf The ByteBuf to which the coordinates are written
+     */
+    public void write(ByteBuf byteBuf) {
+        byteBuf.writeFloat(x);
+        byteBuf.writeFloat(y);
+    }
+
+    /**
+     * Reads the x and y coordinates from the given `byteBuf` and assigns them to the coordinate of this vector.
+     *
+     * @param byteBuf The ByteBuf from which the coordinates are read
+     */
+    public void read(ByteBuf byteBuf) {
+        x = byteBuf.readFloat();
+        y = byteBuf.readFloat();
     }
 }