+++ Changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project follows to [Ragnarök Versioning Convention](https://shor.cz/ragnarok_versioning_convention).
 
+## [Unreleased] Red Core Version 0.3.2 Changelog
+
+## Added
+
+- Added `read` and `write` methods to all vectors that allows for easy networking
+
+---
+
 ## Red Core Version 0.3.1 Changelog - 2023-7-16 [YANKED]
 
 ### Fixed